NEWRELIC-5601: rename messaging application to Nifty Messages for clarity
Hopefully this will make it clearer that we're referring to a
specific, toy messaging application.
